Division Tests por clases
Los test pasaron de estar divididos por iteración a estar divididos por clases
