+++ TP_Colectivo_test/TestTarjeta.cs
@@ -1,10 +1,13 @@
-using NUnit.Framework;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using TP_Colectivo;
 
 namespace TP_Colectivo_test
 {
-    public class Tests
-
+    internal class TestTarjeta
     {
         private Colectivo q;
         private Tarjeta tarjeta;
@@ -19,8 +22,6 @@ namespace TP_Colectivo_test
             q = new Colectivo("Q");
             tarifa = q.VerTarifa();
         }
-
-        [Test]
         public void chequeo_saldo()
         {
             tarjeta.CargarTarjeta(2000);
@@ -67,7 +68,7 @@ namespace TP_Colectivo_test
         public void chequeo_saldonegativo()
         {
             tarjeta.CargarTarjeta(5000);
-            q.pagarCon(tarjeta,tiempo);
+            q.pagarCon(tarjeta, tiempo);
             Assert.That(tarjeta.VerSaldo, Is.EqualTo(5000 - tarifa));
 
             tarjeta.setear(-480 + q.VerTarifa());
@@ -79,6 +80,36 @@ namespace TP_Colectivo_test
             tarjeta.setear(0);
             q.pagarCon(tarjeta, tiempo);
             Assert.That(tarjeta.VerSaldo, Is.EqualTo(0));
+        }
+
+        [Test]
+        public void SaldoDeLaTarjeta()
+        {
+            tarjeta.setear(35000);
+            tarjeta.CargarTarjeta(4000);
+
+
+            Assert.That(tarjeta.VerSaldo, Is.EqualTo(36000));
+            Assert.That(tarjeta.acreditacionPendiente, Is.EqualTo(39000 - 36000));
+
+            q.pagarCon(tarjeta, tiempo);
+            Assert.That(tarjeta.acreditacionPendiente, Is.EqualTo(3000 - tarifa));
+
+            q.pagarCon(tarjeta, tiempo);
+            q.pagarCon(tarjeta, tiempo);
+
+            Assert.That(tarjeta.acreditacionPendiente, Is.EqualTo(0));
+
+            tarjeta.CargarTarjeta(2000);
+
+            Assert.Pass();
+        }
+
+        [Test]
+        public void checkeomes()
+        {
+            tarjeta.setviajesmes(88);
+            Assert.That(tarjeta.ViajesMes, Is.EqualTo(88));
         }
     }
 }
